packages/workflows/ingestion/pyproject.toml
[tool.coverage.run]
branch = true
source = [ "ingestion_workflow" ]

[tool.coverage.report]
exclude_lines = [ "if TYPE_CHECKING:" ]
show_missing = true

[tool.pytest.ini_options]
addopts = """
 --cov
 --cov-report html:'../../../coverage/packages/workflows/ingestion/html'
 --cov-report xml:'../../../coverage/packages/workflows/ingestion/coverage.xml'
 --html='../../../reports/packages/workflows/ingestion/unittests/html/index.html'
 --junitxml='../../../reports/packages/workflows/ingestion/unittests/junit.xml'
"""

[tool.hatch.build.targets.wheel]
packages = [ "ingestion_workflow" ]

[tool.hatch.metadata]
allow-direct-references = true

[tool.ruff]
exclude = [ ".ruff_cache", ".svn", ".tox", ".venv", "dist" ]
line-length = 88
indent-width = 4

  [tool.ruff.lint]
  select = [ "E", "F", "UP", "B", "SIM", "I" ]
  ignore = [ "B008" ]
  fixable = [ "ALL" ]
  unfixable = [ ]

[tool.uv.sources]
utils = { workspace = true }

[project]
name = "ingestion-workflow"
version = "1.0.0"
description = "Automatically generated by Nx."
requires-python = ">=3.9,<4"
readme = "README.md"
dependencies = [
 "aioboto3>=14.3.0",
 "alembic>=1.15.2",
 "dependency-injector>=4.46.0",
 "fastapi>=0.115.12",
 "langchain-community>=0.3.24",
 "langchain-openai>=0.3.16",
 "langchain-text-splitters>=0.3.8",
 "numpydantic>=1.6.9",
 "pgvector>=0.4.1",
 "psycopg[binary]>=3.2.8",
 "pypdf>=5.5.0",
 "sqlmodel>=0.0.24",
 "temporalio>=1.11.1",
 "tiktoken>=0.9.0",
 "utils",
 "uvicorn>=0.34.2",
 "vm-x-ai-sdk==1.6.0",
]

[build-system]
requires = [ "hatchling" ]
build-backend = "hatchling.build"

[dependency-groups]
dev = [
    "types-aioboto3[s3]>=14.3.0",
    "types-aiobotocore[s3]>=2.22.0",
    "watchdog>=6.0.0",
]
